add js file and make refactoring
